# app/OpenApi/Schemas/BrandModel.php
<?php

namespace App\OpenApi\Schemas;

use OpenApi\Annotations as OA;

/**
 * @OA\Schema(
 *     schema="BrandModel",
 *     type="object",
 *     title="Brand",
 *
 *     @OA\Property(property="id", type="integer", example=1),
 *     @OA\Property(property="name", type="string", example="Lada"),
 * )
 */
class BrandModel {}